% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttributes.R
\name{getAttributes}
\alias{getAttributes}
\title{getAttributes}
\usage{
getAttributes(database, mart, filters, value, cond)
}
\arguments{
\item{database}{RSQLite object imported previously to R}

\item{mart}{Dataset name, written inside ""}

\item{filters}{A character vector with restrictions of the query}

\item{value}{A character vector with the corresponding restriction values}

\item{cond}{Conditional when more than one filter is used. Can be "and", "or","not"
#' @examples
regulon <- importDB("~/Documents/regulondb_sqlite3.db")
getAttr(
database = regulon,
mart = "GENE",
filters = c("strand", "celullar_location"),
value = c("forward","cytosol" ),
cond = "and")}
}
\description{
Allows users to filter and retrieve data from a particular mart.
}
