% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegulatedGenesByTF.R
\name{getRegulatedGenesByTF}
\alias{getRegulatedGenesByTF}
\title{getRegulatedGenesByTF}
\usage{
getRegulatedGenesByTF(database, tf)
}
\arguments{
\item{database}{RSQLite object imported previously from an SQLite database file to R}

\item{mart}{Dataset table into the database, use the syntax "dataset"}
}
\description{
Shows all the genes regulated by a Transcription Factor indicated by the user
}
\examples{
## import the SQLite database
## path_sqlite3 <- file.choose()
regulon <- DBI::dbConnect(RSQLite::SQLite(), path_sqlite3)
listDatasets(regulon)
listAttributes(regulon, "TF")
getAttr(
database = regulon,
mart = "GENE",
filters = c("operon_id", "operon_name"),
value = c("ECK125235983","cmoM-mukFEB" ),
cond = "and")
GetRegulatedGenesByTF(regulon, "Ara")
 dbDisconnect(regulon)
}
