% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDB.R
\name{importDB}
\alias{importDB}
\title{importDB}
\usage{
importDB(file)
}
\arguments{
\item{file}{local path of the SQL file, written inside ""}
}
\description{
Import an SQLite database from a local path
}
\examples{
regulon <- importDB("~/Documentsregulondb_sqlite3.db")
}
