% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAttributes.R
\name{listAttributes}
\alias{listAttributes}
\title{listAttributes}
\usage{
listAttributes(database, mart)
}
\arguments{
\item{database}{RSQLite object imported previously to R}

\item{mart}{Dataset name, written inside ""}
}
\description{
Prints the column names of the mart
}
\examples{
regulon <- importDB("~/Documents/regulondb_sqlite3.db")
listAttributes(regulon, "TF")
}
