% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDatasets.R
\name{listDatasets}
\alias{listDatasets}
\title{listDatasets}
\usage{
listDatasets(database)
}
\arguments{
\item{database}{RSQLite object imported previously to R}
}
\description{
Prints all marts contained in the SQL object
}
\examples{
regulon <- importDB("~/Documentsregulondb_sqlite3.db")
listDatasets(regulon)
}
